+++ apps/pie-storybook/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 1.87.1
+
+### Patch Changes
+
+- Updated dependencies [[`ce29f7a`](https://github.com/justeattakeaway/pie/commit/ce29f7ae238fb5f6b47c4d5ccd86e199138fc728)]:
+  - @justeat/pie-design-tokens@7.14.3
+  - @justeattakeaway/pie-css@1.3.2
+  - @justeattakeaway/pie-webc@0.11.2
+
+## 1.87.0
+
+### Minor Changes
+
+- [Added] - Stories using pie-checkbox in pie-list-item ([#3053](https://github.com/justeattakeaway/pie/pull/3053)) by [@jamieomaguire](https://github.com/jamieomaguire)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-webc@0.11.1
+
 ## 1.86.4
 
 ### Patch Changes